Moved all non define function to globalhelper
